Fix Execution context was destroyed error in e2e tests
